Started working on chat widget. Not fully functional
Added a PreInputProcessor because the txt UMG boxes were stealing all of the input for the Enter key and they didn't have an event that I could use.

Now when you want the player controller to handle enter and escape events you can use bChatCanHandleEnter = true; or bChatCanHandleEscape = true; and they are both exposed to blueprint as well.